third_party/WebKit/LayoutTests/fast/dom/custom/document-register-on-create-callback.html

<!DOCTYPE html>
<script src="../../../resources/testharness.js"></script>
<script src="../../../resources/testharnessreport.js"></script>
<script>
(function() {

var test = async_test('document.registerElement() in a createdCallback should upgrade elements that are in the same processing queue.');

var createdElements = [];

var protoA = Object.create(HTMLElement.prototype);
protoA.createdCallback = function() {
    createdElements.push(this.id);
    var protoB = Object.create(HTMLElement.prototype);
    protoB.createdCallback = function () {
        createdElements.push(this.id);
    };

    document.registerElement('x-b', {prototype: protoB});
    test.step(function() {
        assert_array_equals(createdElements, ['elemA0', 'elemB0', 'elemB1']);
        test.done();
    });
};

document.registerElement('x-a', {prototype: protoA});

})();
</script>
<x-b id='elemB0'></x-b>
<x-a id='elemA0'></x-a>
<x-b id='elemB1'></x-b>
